Add wrappers to allow CSS nonce to be properly used
